# docs/openapi.yaml
openapi: 3.0.3
info:
  title: WiseLabz REST API
  version: 0.1.0
  description: >
    Contract-first OpenAPI spec for the WiseLabz frontend. This is the source of
    truth the frontend generates against (orval) and mocks against (MSW) BEFORE the
    Go backend exists. The backend implements *to* this spec. Keep this file and
    docs/WS_CONTRACT.md in sync with each other.
servers:
  - url: /api
    description: Same-origin API (Vite proxies /api to the Go backend in dev)
  - url: /api/v1
    description: >
      Versioned alias of the same API. Every path is served identically under
      /api and /api/v1; /api is kept for existing clients.
...
security:
  - bearerAuth: []
paths:
...
  /runbooks/{runbookId}/runs:
    parameters:
      - { $ref: '#/components/parameters/RunbookId' }
    get:
      tags: [runbooks]
      operationId: listRunbookRuns
      summary: Run history, newest first with frozen redacted steps
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Paginated run history; viewers can read, API-key restrictions redact steps
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunbookRunPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /runbook-runs/{runId}:
    parameters:
      - { $ref: '#/components/parameters/RunbookRunId' }
    get:
      tags: [runbooks]
      operationId: getRunbookRun
      summary: Read a run, including after its runbook was deleted
      responses:
        '200':
          description: Frozen steps and outcomes; hidden steps omit all authored metadata and error text
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunbookRun' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: In-memory access token (§8.7). Refresh token rides an HttpOnly cookie.
  parameters:
...
    RunbookRunId: { name: runId, in: path, required: true, schema: { type: string } }
    RunbookId: { name: runbookId, in: path, required: true, schema: { type: string } }
...
    Page: { name: page, in: query, required: false, schema: { type: integer, default: 1, minimum: 1 } }
    PageSize: { name: pageSize, in: query, required: false, schema: { type: integer, default: 25, minimum: 1, maximum: 100 } }
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: array
          description: Field-level validation failures, when the error is field-specific.
          items: { $ref: '#/components/schemas/FieldError' }
    FieldError:
      type: object
      required: [field, msg]
      properties:
        field: { type: string, description: "Request field the failure applies to: a body field, a query parameter, or an indexed path such as [0].provider" }
        msg: { type: string, description: Human-readable reason the field was rejected }

    # ── auth ──
...
    RunbookRunState:
      type: string
      enum: [running, waiting_manual, failed, succeeded, cancelled, expired]
    RunbookRunStepState:
      type: string
      enum: [pending, running, waiting, succeeded, failed, skipped, unknown]
    RunbookRunRecord:
      type: object
      required: [id, runbookTitle, state, startedBy, startedAt, updatedAt]
      properties:
        id: { type: string }
        runbookId: { type: string, description: Omitted after runbook deletion }
        runbookTitle: { type: string, description: Title frozen at start }
        state: { $ref: '#/components/schemas/RunbookRunState' }
        reason: { type: string }
        startedBy: { type: string }
        resumedBy: { type: string }
        cancelledBy: { type: string }
        startedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
    RunbookRunStep:
      type: object
      description: >
        Frozen step and its current outcome. A redacted step carries no kind, title, connector,
        entity, verb, timeout, error or preview; only identifiers, position, state, times,
        confirming user and permission flags remain. Manual steps have no connector and are visible.
      required: [id, position, redacted, canExecute]
      properties:
        id: { type: string }
        position: { type: integer }
        kind: { type: string, enum: [lifecycle, sync_and_wait, wait_until_healthy, manual] }
        title: { type: string }
        connectorId: { type: string }
        connectorName: { type: string }
        verb: { type: string, enum: [restart, start, stop] }
        entityRef: { type: string }
        timeoutSeconds: { type: integer }
        state: { $ref: '#/components/schemas/RunbookRunStepState' }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
        error: { type: string }
        confirmedBy: { type: string }
        redacted: { type: boolean }
        canExecute: { type: boolean }
        executeBlockedReason: { type: string }
        preview: { $ref: '#/components/schemas/RestartPreview' }
    RunbookRun:
      allOf:
        - { $ref: '#/components/schemas/RunbookRunRecord' }
        - type: object
          required: [steps]
          properties:
            steps:
              type: array
              items: { $ref: '#/components/schemas/RunbookRunStep' }
    RunbookRunPage:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/RunbookRun' }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...
    RestartPreview:
      type: object
      description: Non-mutating preview of the future service.restart action.
      required: [targetService, estimatedDowntimeSeconds, dependentServices]
      properties:
        targetService: { type: string }
        estimatedDowntimeSeconds: { type: integer, minimum: 0, example: 30 }
        affectedEntities:
          type: array
          description: Names of devices the target powers or carries (e.g. the APs and cameras behind a UniFi PoE port); the blast radius of the action. Empty when the target declares none.
          items: { type: string }
        dependentServices:
          type: array
          items:
            type: object
            required: [kind, name]
            properties:
              kind: { type: string, description: 'e.g. host, network, storage, upstream_service' }
              name: { type: string }
              ref: { type: string }
...